Move DeprecationWarning to Helper
